Move remaining spec to CLI tests
Mock $.when like for deleteLayer, to avoid duplicated test files
